Relicense and reorganize project
Relicense the project to Apache License 2.0 as used by the LazyVim
repo that this one originated from. Then relocate the configuration
files from the .config/nvim directory to the root of the repository.
I'm doing this so I can remove the reliance on "stow" to link this
from a seperate directory to my ~/.config/nvim directory. I'll just
check it out there directly.